+++ src/com/oDesk/api/Routers/Mc.java
@@ -60,7 +60,61 @@ public final class Mc {
     public JSONObject getThreadDetails(String username, String threadId) throws JSONException {
         return oClient.get("/mc/v1/threads/" + username + "/" + threadId);
     }
-	
+    
+    /**
+     * Get a specific thread by context
+     *
+     * @param   username Username
+     * @param	jobKey Job key
+     * @param   applicationId Application ID
+     * @throws	JSONException If error occurred
+	 * @return	{@link JSONObject}
+     */
+    public JSONObject getThreadByContext(String username, String jobKey, String applicationId) throws JSONException {
+        return oClient.get("/mc/v1/contexts/" + username + "/Interviews:" + jobKey + ":" + applicationId);
+    }
+    
+    /**
+     * Get a specific thread by context
+     *
+     * @param   username Username
+     * @param	jobKey Job key
+     * @param   applicationId Application ID
+     * @param	context Context
+     * @throws	JSONException If error occurred
+	 * @return	{@link JSONObject}
+     */
+    public JSONObject getThreadByContext(String username, String jobKey, String applicationId, String context) throws JSONException {
+        return oClient.get("/mc/v1/contexts/" + username + "/" + context + ":" + jobKey + ":" + applicationId);
+    }
+    
+    /**
+     * Get a specific thread by context (last message content)
+     *
+     * @param   username Username
+     * @param	jobKey Job key
+     * @param   applicationId Application ID
+     * @throws	JSONException If error occurred
+	 * @return	{@link JSONObject}
+     */
+    public JSONObject getThreadByContextLastPosts(String username, String jobKey, String applicationId) throws JSONException {
+        return oClient.get("/mc/v1/contexts/" + username + "/Interviews:" + jobKey + ":" + applicationId + "/last_posts");
+    }
+    
+    /**
+     * Get a specific thread by context (last message content)
+     *
+     * @param   username Username
+     * @param	jobKey Job key
+     * @param   applicationId Application ID
+     * @param	context Context
+     * @throws	JSONException If error occurred
+	 * @return	{@link JSONObject}
+     */
+    public JSONObject getThreadByContextLastPosts(String username, String jobKey, String applicationId, String context) throws JSONException {
+        return oClient.get("/mc/v1/contexts/" + username + "/" + context + ":" + jobKey + ":" + applicationId + "/last_posts");
+    }
+    
     /**
      * Send new message
      *